feat: 세일 alert 함수 분리
